Improve gnunet-bcd and update logo used in generated cards
The HTML forms now use a custom CSS instead of importing Bootstrap 4. This new
CSS is basd on Bootstrap 5 with some minimal differences to make it look a bit
like the older theme.

Bootstrap was removed because these pages don't need its full power and also
because it was difficult for some text editors to handle the one-long-line
minified CSS embedded in the HTML.  Since the forms are just a series of label
+ input elements, using a specific set of directives doesn't add any
significant maintenance cost while making editing easiers.

A new form to generate a simplified card was added.  This card is desgined to
reduce the leftover whitespace when some optional entries are left blank in
the "full" form, while at the same time giving importance to the QR code
containing the GNS address.  Its layout can be improved.

The simplified form can also be used to generate a PNG picture of the QR code,
so that it can be included in other graphical media other than a business
card, for example on a flyer or in the signature of a HTML-based e-mail.

Generally, the gnunet-bcd service should provide a slightly better user
experience, even if it's just to provide different pages for different errors
but with a unified style.

The logo used in the cards has been changed from the previous "gnu in front of
a spiderweb" illustration to the new "gnu-shaped network of peers" image.  It
is also now a pure TikZ picture instead of being made using PSTricks, because,
apparently, there is an incompatibility between the two packages in recent
versions which made it impossible to compile the file.

As a consequence of the incompatibility between TikZ and PSTricks, the QR code
is now generated using the 'qrcode' package.

This commit contains some changes to FCFS, but they're just some minor fixes
and are unrelated to the changes to BCD.