-fix subtle bug in GNUNET_CONFIGURATION_default
Apparently this was there since the beginning and it wasn't caught earlier
merely due to a coincidence.

Basically, it was looking at the caller's values instead of GNUnet's and
even when I used this function in personal projects, it just happened that I
was calling this function before setting the new project data, so the two
environments ended up being the same.

It didn't cause any issues because it was still returning GNUnet's own
configuration, meaning everything else worked as expected, but naturally if
one were to move the call later on it would break.

Also add a comment to answer the FIXME.