Update configure.ac and a few macros used by it
The changes done to configure.ac are essentially divided in two categories:
the first is made of changes to improve existing checks, even if it's merely
in the message being printed during the check; the second category is made of
updates aimed at bringing this file closer to Autoconf 2.71, which deprecated
or even removed some stuff that was otherwise present either directly in
configure.ac or in one of the macros it includes, which have been updated in
this commit too.

Actually, the generated configure script was created by Autoconf 2.71 during
deveopment, so in a way it's already compatible with the latest version.