configure.ac: use lo0 for the 3 listed BSD Systems.
Without this, GNUNET_DEFAULT_INTERFACE is empty, which leads to a
compilation-time error on at least NetBSD due to the resulting
syntax error.

lo0 definitely exists, it is up to the individual package maintainers and/or
system administrators after installation to set this to usable devices. We can
not predict the default device names on those *BSD systems, as there are no
default names.

Signed-off-by: Martin Schanzenbach <schanzen@gnunet.org>