TESTING: The testing API should properly take strings as const
Strings passed to the API should either be const and copied, or handled
by the caller and free'd. Looking at the transport tests currently
neither is happening. The strings are probably safer but definitely
leaked.
Changed the API to use "const" strings. Users of the testing API must
define functions that copy (and free!) strings accordingly.