UTIL: add test demonstrating scheduler bug, don't run it by default.
These demonstrate a bug in the scheduler by which a task can prevent any other
task from running for an arbitrarily long time despite regularly yielding to
the scheduler.  It is caused by a faulty check in GNUNET_SCHEDULER_do_work
that assumes that the task that was the last in the queue when the pass began
will still be in the same relative position when the pass ends, and uses this
assumption to detect the end of the current pass.  This assumption fails when
the last task of the current pass is canceled after the pass has started.  It
also fails when we schedule a higher-priority task to run immediately, causing
work_priority to immediately switch such that we now process a queue that
doesn't contain the pass-ending task we're looking for.

These tests are built, but not run by 'make check' yet, since they currently
fail.  You can manually verify that they do currently fail.

Signed-off-by: Christian Grothoff <christian@grothoff.org>