UTIL: use dedicated marker in ready queue.
This inserts a dedicated dummy marker task at the end of the ready queue at
the start of a pass.  Because this marker task isn't visible to users of the
scheduler, it can't be canceled while the pass is being run.  Additionally,
switching which ready queue is being run partway through by scheduling a
higher-priority task to immediately run also places this dummy marker.  This
resolves both erroneous cases by which a pass can accidentally run an
unbounded number of tasks.

This also modifies GNUNET_SCHEDULER_get_load to not be misled by this extra
dummy task, and adds the now-passing test cases to the test suite.

Signed-off-by: Christian Grothoff <christian@grothoff.org>