SETU: avoid 64-bit shift on 64-bit value.
Shifting a 64-bit value by any more than 63 bits is undefined behavior,
apparently - at least, the sanitizers complain about it. The intuitive,
obvious result, of course, is for the result to be 0. In this case, when s ==
0, x << (64 - s) should result in 0, and (x >> s) should result in x, and the
bitwise-or of those two should be x. Which x already was.

Perhaps it should be investigated whether (x >> (64 - s)) should actually
be (x >> (63 - s)), since 0 <= s < 64.

Signed-off-by: Martin Schanzenbach <schanzen@gnunet.org>