REVOCATION: add shutdown task earlier.
There are many things that can go wrong and require cleanup before the
shutdown task is currently added. In these cases, GNUNET_SCHEDULER_shutdown is
run and run() is returned from. At present, the only harm in running
shutdown_task earlier would be it unconditionally destroying revocation_map,
so we can add it as soon as that exists.

Signed-off-by: Martin Schanzenbach <schanzen@gnunet.org>