NAMESTORE: fix overread in handle_record_store.
A RecordStoreMessage looks like this:

| header | key | recordset |

A StoreActivity's rs field is supposed to point to the record
set. handle_record_store tries to make a copy of this record set, but it does
it by allocating enough memory for both key and recordset, then copying
sizeof(key) + sizeof(recordset) bytes into it *starting from recordset*. This
causes memcpy to read past the end of recordset by sizeof(key) bytes. There's
still enough room in the allocated region for it, though, so it's only an
overread.

Signed-off-by: Martin Schanzenbach <schanzen@gnunet.org>