TRANSPORT: fix contains_address.
GNUNET_memcmp takes two typed pointers and compares the bytes of the objects
pointed to, using the size of the type of the first argument. The first
argument here is of type char *, so it compares the character it points to to
the character the second argument points to. This is probably not what was
intended.

Signed-off-by: Martin Schanzenbach <schanzen@gnunet.org>