pq: make oid retrieval schema aware
The previous way we retrieved the OID for a given typename
would retrieve all entries (accross all schemas) in the
pg_type table, but limit the result to 1.

This led to consistency check errors in the following
scenario: After array data was written to a table
(implicitly marked as, say, type oid_1), when a new type of
the same name was added (in a different schema), _their_
oid_2 was returned for the typename, and the consistency
check failed.

This commit simplifies the select statement we need for the
oid retrieval and at the same time ensures we receive the
oid of a type in the _current_ schema of the caller.